requirements.txt
Django==1.9
Flask==0.10.1
django-bitfield==1.8.0
django-extensions==1.6.1
djangorestframework==3.3.2
logging==0.4.9.6
lxml==3.4.0
matplotlib==1.4.0
mysql-connector-python==2.0.2
numpy==1.9.0
pyreadline==2.0
pyserial==2.7
rollbar==0.11.0
scipy==0.14.0